mavenbasico/src/main/java/com/arquitecturajava/proyecto1/models/Persona.java
package com.arquitecturajava.proyecto1.models;

public class Persona {

	private String nombre;
	private int edad;
	public String getNombre() {
		return nombre;
	}
	public void setNombre(String nombre) {
		this.nombre = nombre;
	}
	public int getEdad() {
		return edad;
	}
	public void setEdad(int edad) {
		this.edad = edad;
	}
	public Persona(String nombre, int edad) {
		super();
		this.nombre = nombre;
		this.edad = edad;
	}
	
	public boolean estaJubilado() {
		
		return edad>65;
		
	}
	
	
}
